+++ src/Bootstrap.php
@@ -3,9 +3,12 @@
 namespace davidhirtz\yii2\cms;
 
 use davidhirtz\yii2\cms\models\Asset;
+use davidhirtz\yii2\cms\models\events\FileBeforeDeleteEventHandler;
+use davidhirtz\yii2\media\models\File;
 use davidhirtz\yii2\skeleton\web\Application;
 use Yii;
 use yii\base\BootstrapInterface;
+use yii\base\ModelEvent;
 use yii\i18n\PhpMessageSource;
 
 class Bootstrap implements BootstrapInterface
@@ -39,13 +42,20 @@ class Bootstrap implements BootstrapInterface
             ],
             'media' => [
                 'class' => \davidhirtz\yii2\media\Module::class,
-                'assets' => [
-                    Asset::class,
-                ],
+                'fileRelations' => [Asset::class],
             ],
         ]);
 
         $this->addDefaultUrlRules();
+
+        ModelEvent::on(
+            File::class,
+            File::EVENT_BEFORE_DELETE,
+            fn (ModelEvent $event) => Yii::createObject(FileBeforeDeleteEventHandler::class, [
+                $event,
+                $event->sender,
+            ])
+        );
 
         $app->setMigrationNamespace('davidhirtz\yii2\cms\migrations');
     }